// src/app/backend/models/players.ts
export interface Players {
    id: number;
    name: string;
    hp: number;
    attack: number;
    defense: number;
    experience: number;
    level: number;
    honors: number;
    money: number;
    job: string;
    skill1Name: string;
    skill1Description: string;
    skill1Cooldown: number;
    skill1LvlUnlock: number;
    skill2Name: string;
    skill2Description: string;
    skill2Cooldown: number;
    skill2LvlUnlock: number;
    skill3Name: string;
    skill3Description: string;
    skill3Cooldown: number;
    skill3LvlUnlock: number;
    uniqueAbilityName: string;
    uniqueAbilityDescription: string;
    uniqueAbilityCooldown: number;
    lider: boolean;
}